Update UserService to allow cancelled logins
Fixes #51

UserService was caching the deferred prematurely (without waiting for the promise to be resolved) which was causing the UserService to act as though the user was defined even if the user was never retrieved from the service.  Now the deferred is only cached when the service actually retrieves a success response from the API server.